Refactor server handlers into SRP modules
