+++ crates/hl7v2-server/src/handlers.rs
@@ -3,17 +3,31 @@
 use axum::{
     extract::{Json, State},
     http::StatusCode,
-    response::{IntoResponse, Response},
+    response::IntoResponse,
 };
-use sha2::{Digest, Sha256};
-use std::collections::BTreeMap;
 use std::sync::Arc;
 
-use crate::PROFILE_LOAD_SAFE_MESSAGE;
 use crate::audit::{self, MessageLogContext};
 use crate::models::*;
-use crate::redaction::redact_message;
 use crate::server::AppState;
+
+mod ack_policy;
+mod corpus;
+mod error;
+mod metadata;
+mod parsing;
+mod quarantine;
+mod schema_versions;
+
+use ack_policy::*;
+use corpus::*;
+pub use error::AppError;
+use metadata::extract_metadata;
+#[cfg(test)]
+use metadata::joined_components;
+use parsing::*;
+use quarantine::*;
+use schema_versions::*;
 
 /// Handler for GET /health
 pub async fn health_handler(State(state): State<Arc<AppState>>) -> impl IntoResponse {
@@ -662,757 +676,6 @@ pub async fn normalize_handler(
     );
 
     Ok((StatusCode::OK, Json(response)))
-}
-
-fn parse_request_message(
-    message_bytes: &[u8],
-    mllp_framed: bool,
-) -> Result<hl7v2::Message, AppError> {
-    if mllp_framed {
-        hl7v2::parse_mllp(message_bytes)
-            .map_err(|e| AppError::Parse(format!("MLLP parse error: {}", e)))
-    } else {
-        hl7v2::parse(message_bytes).map_err(|e| AppError::Parse(format!("Parse error: {}", e)))
-    }
-}
-
-fn parse_request_message_with_metrics(
-    message_bytes: &[u8],
-    mllp_framed: bool,
-    operation: &'static str,
-) -> Result<hl7v2::Message, AppError> {
-    match parse_request_message(message_bytes, mllp_framed) {
-        Ok(message) => {
-            crate::metrics::record_parse_success(operation, message_bytes.len());
-            Ok(message)
-        }
-        Err(error) => {
-            crate::metrics::record_parse_failure(operation);
-            Err(error)
-        }
-    }
-}
-
-fn redact_message_with_metrics(
-    message: &mut hl7v2::Message,
-    policy_toml: &str,
-    operation: &'static str,
-) -> Result<RedactionReceipt, AppError> {
-    match redact_message(message, policy_toml) {
-        Ok(receipt) => Ok(receipt),
-        Err(error) => {
-            crate::metrics::record_redaction_failure(operation);
-            Err(AppError::Redaction(error))
-        }
-    }
-}
-
-fn requested_report_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported validation report schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_redaction_receipt_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported redaction receipt schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_quarantine_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported quarantine output schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_bundle_artifact_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported bundle artifact schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_replay_report_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported replay report schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_corpus_summary_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported corpus summary schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_corpus_fingerprint_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported corpus fingerprint schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn requested_corpus_diff_schema_version(version: Option<u8>) -> Result<u8, AppError> {
-    match version.unwrap_or(1) {
-        1 => Ok(1),
-        2 => Ok(2),
-        other => Err(AppError::Validation(format!(
-            "unsupported corpus diff schema version {other}; expected 1 or 2"
-        ))),
-    }
-}
-
-fn validated_corpus_message_ids(
-    messages: &[CorpusMessageInput],
-    field_name: &str,
-    default_prefix: &str,
-) -> Result<Vec<String>, AppError> {
-    if messages.is_empty() {
-        return Err(AppError::Validation(format!(
-            "{field_name} must contain at least one message"
-        )));
-    }
-
-    messages
-        .iter()
-        .enumerate()
-        .map(|(index, message)| {
-            let label = message
-                .id
-                .clone()
-                .unwrap_or_else(|| format!("{default_prefix}-{}", index.saturating_add(1)));
-            validate_corpus_message_id(&label)?;
-            Ok(label)
-        })
-        .collect()
-}
-
-fn validate_corpus_message_id(label: &str) -> Result<(), AppError> {
-    if label.is_empty() || label == "." || label == ".." || label.len() > 128 {
-        return Err(AppError::Validation(
-            "corpus message id must be 1-128 characters and cannot be '.' or '..'".to_string(),
-        ));
-    }
-
-    if !label
-        .chars()
-        .all(|character| character.is_ascii_alphanumeric() || matches!(character, '.' | '_' | '-'))
-    {
-        return Err(AppError::Validation(
-            "corpus message id must use only ASCII letters, numbers, '.', '_' or '-'".to_string(),
-        ));
-    }
-
-    Ok(())
-}
-
-fn corpus_message_refs<'a>(
-    messages: &'a [CorpusMessageInput],
-    ids: &'a [String],
-) -> Vec<hl7v2::synthetic::corpus::CorpusMessageRef<'a>> {
-    messages
-        .iter()
-        .zip(ids.iter())
-        .map(|(message, id)| {
-            hl7v2::synthetic::corpus::CorpusMessageRef::new(id.as_str(), message.message.as_bytes())
-        })
-        .collect()
-}
-
-fn attach_profile_to_fingerprint(
-    fingerprint: &mut hl7v2::synthetic::corpus::CorpusFingerprint,
-    profile_yaml: &str,
-    messages: &[CorpusMessageInput],
-) -> Result<hl7v2::synthetic::corpus::CorpusFingerprintProfile, AppError> {
-    let profile = hl7v2::load_profile_checked(profile_yaml).map_err(AppError::from)?;
-    let metadata = hl7v2::synthetic::corpus::CorpusFingerprintProfile {
-        path: "<inline-profile>".to_string(),
-        sha256: compute_sha256(profile_yaml),
-        version: profile.version.clone(),
-        message_structure: profile.message_structure.clone(),
-    };
-    fingerprint.profile = Some(metadata.clone());
-    fingerprint.validation_issue_code_counts =
-        validation_issue_counts_for_loaded_profile(messages, &profile);
-    Ok(metadata)
-}
-
-fn validation_issue_counts_for_messages(
-    messages: &[CorpusMessageInput],
-    profile_yaml: &str,
-) -> Result<Vec<hl7v2::synthetic::corpus::CorpusCount>, AppError> {
-    let profile = hl7v2::load_profile_checked(profile_yaml).map_err(AppError::from)?;
-    Ok(validation_issue_counts_for_loaded_profile(
-        messages, &profile,
-    ))
-}
-
-fn validation_issue_counts_for_loaded_profile(
-    messages: &[CorpusMessageInput],
-    profile: &hl7v2::Profile,
-) -> Vec<hl7v2::synthetic::corpus::CorpusCount> {
-    let mut counts: BTreeMap<String, usize> = BTreeMap::new();
-
-    for message in messages {
-        let parsed = if hl7v2::is_mllp_framed(message.message.as_bytes()) {
-            hl7v2::parse_mllp(message.message.as_bytes())
-        } else {
-            hl7v2::parse(message.message.as_bytes())
-        };
-        let Ok(parsed) = parsed else {
-            continue;
-        };
-
-        let issues = hl7v2::validate(&parsed, profile);
-        let report = hl7v2::ValidationReport::from_issues(
-            &parsed,
-            Some(profile.message_structure.clone()),
-            issues,
-        );
-        for issue in report.issues {
-            let count = counts.entry(issue.code).or_insert(0);
-            *count = count.saturating_add(1);
-        }
-    }
-
-    counts
-        .into_iter()
-        .map(|(value, count)| hl7v2::synthetic::corpus::CorpusCount { value, count })
-        .collect()
-}
-
-fn validation_report_v2_for_server(
-    report: &hl7v2::ValidationReport,
-    profile_yaml: &str,
-    profile: &hl7v2::Profile,
-) -> hl7v2::ValidationReportV2 {
-    report.to_v2(
-        "hl7v2-server",
-        env!("CARGO_PKG_VERSION"),
-        Some(hl7v2::ValidationReportProfileIdentity {
-            label: profile.message_structure.clone(),
-            message_structure: Some(profile.message_structure.clone()),
-            version: Some(profile.version.clone()),
-            sha256: Some(compute_sha256(profile_yaml)),
-        }),
-    )
-}
-
-fn compute_sha256(value: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(value.as_bytes());
-    format!("{:x}", hasher.finalize())
-}
-
-fn parse_msh_for_ack_policy(
-    message_bytes: &[u8],
-    mllp_framed: bool,
-) -> Result<hl7v2::Message, AppError> {
-    let input = if mllp_framed {
-        hl7v2::unwrap_mllp(message_bytes)
-            .map_err(|e| AppError::Parse(format!("MLLP parse error: {}", e)))?
-    } else {
-        message_bytes
-    };
-
-    let first_segment_end = input
-        .iter()
-        .position(|byte| matches!(byte, b'\r' | b'\n'))
-        .unwrap_or(input.len());
-    let msh = &input[..first_segment_end];
-    if !msh.starts_with(b"MSH") {
-        return Err(AppError::Parse(
-            "Parse error: message did not contain a usable MSH segment".to_string(),
-        ));
-    }
-
-    let mut buffer = msh.to_vec();
-    buffer.push(b'\r');
-    hl7v2::parse(&buffer)
-        .map_err(|e| AppError::Parse(format!("MSH parse error for ACK policy: {}", e)))
-}
-
-fn map_ack_code(code: AckRequestCode) -> hl7v2::AckCode {
-    match code {
-        AckRequestCode::Aa => hl7v2::AckCode::AA,
-        AckRequestCode::Ae => hl7v2::AckCode::AE,
-        AckRequestCode::Ar => hl7v2::AckCode::AR,
-        AckRequestCode::Ca => hl7v2::AckCode::CA,
-        AckRequestCode::Ce => hl7v2::AckCode::CE,
-        AckRequestCode::Cr => hl7v2::AckCode::CR,
-    }
-}
-
-fn ack_policy_decision_for_validation(
-    policy: &AckPolicyConfig,
-    report: &hl7v2::ValidationReport,
-) -> Result<AckPolicyDecision, AppError> {
-    if report.valid && policy.accept_on == AckPolicyAcceptOn::Valid {
-        let ack_code = match policy.mode {
-            AckPolicyMode::Original => AckRequestCode::Aa,
-            AckPolicyMode::Enhanced => AckRequestCode::Ca,
-        };
-        return Ok(AckPolicyDecision {
-            mode: policy.mode,
-            outcome: AckPolicyOutcome::Accepted,
-            reason: AckPolicyReason::Valid,
-            ack_code: ack_code.as_str().to_string(),
-            include_error_text: false,
-            error_text: None,
-        });
-    }
-
-    if policy.rejects(AckPolicyRejectCondition::ValidationError) {
-        return Ok(ack_policy_reject_decision(
-            policy,
-            AckPolicyReason::ValidationError,
-            report.issue_count,
-        ));
-    }
-
-    Err(AppError::Validation(
-        "ACK policy did not define a decision for validation failure".to_string(),
-    ))
-}
-
-fn ack_policy_reject_decision(
-    policy: &AckPolicyConfig,
-    reason: AckPolicyReason,
-    issue_count: usize,
-) -> AckPolicyDecision {
-    let ack_code = match policy.mode {
-        AckPolicyMode::Original => AckRequestCode::Ar,
-        AckPolicyMode::Enhanced => AckRequestCode::Cr,
-    };
-    let error_text = policy
-        .include_error_text
-        .then(|| ack_policy_error_text(reason, issue_count));
-
-    AckPolicyDecision {
-        mode: policy.mode,
-        outcome: AckPolicyOutcome::Rejected,
-        reason,
-        ack_code: ack_code.as_str().to_string(),
-        include_error_text: policy.include_error_text,
-        error_text,
-    }
-}
-
-fn ack_policy_error_text(reason: AckPolicyReason, issue_count: usize) -> String {
-    match reason {
-        AckPolicyReason::Valid => "message accepted".to_string(),
-        AckPolicyReason::ParseError => "message parsing failed".to_string(),
-        AckPolicyReason::ValidationError => {
-            format!("message validation failed with {issue_count} issue(s)")
-        }
-    }
-}
-
-fn ack_code_from_policy_decision(decision: &AckPolicyDecision) -> Result<hl7v2::AckCode, AppError> {
-    match decision.ack_code.as_str() {
-        "AA" => Ok(hl7v2::AckCode::AA),
-        "AR" => Ok(hl7v2::AckCode::AR),
-        "CA" => Ok(hl7v2::AckCode::CA),
-        "CR" => Ok(hl7v2::AckCode::CR),
-        code => Err(AppError::Internal(format!(
-            "ACK policy produced unsupported ACK code: {code}"
-        ))),
-    }
-}
-
-struct RedactedQuarantineContext<'a> {
-    state: &'a AppState,
-    raw_input: &'a [u8],
-    profile_yaml: &'a str,
-    policy_text: &'a str,
-    redacted_message: &'a hl7v2::Message,
-    redacted_hl7: &'a str,
-    redaction_receipt: &'a RedactionReceipt,
-    validation_report: &'a hl7v2::ValidationReport,
-}
-
-fn maybe_write_redacted_quarantine(
-    context: RedactedQuarantineContext<'_>,
-) -> Result<Option<QuarantineOutputSummary>, AppError> {
-    let RedactedQuarantineContext {
-        state,
-        raw_input,
-        profile_yaml,
-        policy_text,
-        redacted_message,
-        redacted_hl7,
-        redaction_receipt,
-        validation_report,
-    } = context;
-
-    if validation_report.valid || !state.quarantine.enabled {
-        return Ok(None);
-    }
-
-    let root = state
-        .quarantine
-        .path
-        .as_deref()
-        .ok_or(AppError::QuarantineOutputNotConfigured)?;
-    let output_id = generated_quarantine_id();
-    let summary =
-        crate::evidence::write_quarantine_output(crate::evidence::QuarantineOutputWriteRequest {
-            root,
-            output_id: &output_id,
-            config: &state.quarantine,
-            raw_input,
-            profile_yaml,
-            policy_text,
-            redacted_message,
-            redacted_hl7,
-            redaction_receipt,
-            validation_report,
-        })
-        .map_err(AppError::quarantine_from_evidence_error)?;
-
-    Ok(Some(summary))
-}
-
-fn generated_quarantine_id() -> String {
-    let nanos = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map_or(0, |duration| duration.as_nanos());
-    format!("quarantine-{}-{nanos}", std::process::id())
-}
-
-/// Extract message metadata from parsed message
-fn extract_metadata(message: &hl7v2::Message) -> Result<MessageMetadata, AppError> {
-    // Find MSH segment
-    let msh = message
-        .segments
-        .first()
-        .ok_or_else(|| AppError::Parse("Missing MSH segment".to_string()))?;
-
-    if &msh.id != b"MSH" {
-        return Err(AppError::Parse("First segment must be MSH".to_string()));
-    }
-
-    // Extract MSH fields
-    let message_type = joined_components(message, "MSH.9").unwrap_or_else(|| "UNKNOWN".to_string());
-
-    let version = hl7v2::get(message, "MSH.12").unwrap_or("2.5").to_string();
-
-    let sending_application = hl7v2::get(message, "MSH.3").unwrap_or("").to_string();
-
-    let sending_facility = hl7v2::get(message, "MSH.4").unwrap_or("").to_string();
-
-    let message_control_id = hl7v2::get(message, "MSH.10").unwrap_or("").to_string();
-
-    Ok(MessageMetadata {
-        message_type,
-        version,
-        sending_application,
-        sending_facility,
-        message_control_id,
-        segment_count: message.segments.len(),
-        charsets: message.charsets.clone(),
-    })
-}
-
-fn joined_components(message: &hl7v2::Message, path: &str) -> Option<String> {
-    let mut components = Vec::new();
-
-    for index in 1.. {
-        let component_path = format!("{}.{}", path, index);
-        match hl7v2::get(message, &component_path) {
-            Some(value) if !value.is_empty() => components.push(value.to_string()),
-            Some(_) => {}
-            None => break,
-        }
-    }
-
-    if components.is_empty() {
-        hl7v2::get(message, path).map(str::to_string)
-    } else {
-        Some(components.join("^"))
-    }
-}
-
-/// Application error type with specific error variants.
-///
-/// This enum provides detailed error information for different failure modes,
-/// making it easier to diagnose issues and provide meaningful error responses.
-#[derive(Debug)]
-pub enum AppError {
-    /// Message parsing error (malformed HL7, invalid structure, etc.)
-    Parse(String),
-
-    /// Profile loading error (YAML syntax, missing fields, etc.)
-    ProfileLoad(String),
-
-    /// Validation error (message does not conform to profile)
-    Validation(String),
-
-    /// Redaction policy or redaction application error
-    Redaction(String),
-
-    /// Bundle output is not configured on the server
-    BundleOutputNotConfigured,
-
-    /// Bundle output root is configured but not writable or available
-    BundleOutputNotReady(String),
-
-    /// Bundle request or write error
-    Bundle(String),
-
-    /// Bundle output already exists
-    Conflict(String),
-
-    /// Requested bundle id was not found under the configured output root
-    BundleNotFound(String),
-
-    /// Quarantine output is enabled but no path is configured
-    QuarantineOutputNotConfigured,
-
-    /// Quarantine output root is configured but not writable or available
-    QuarantineOutputNotReady(String),
-
-    /// Quarantine output write error
-    Quarantine(String),
-
-    /// Quarantine output already exists
-    QuarantineConflict(String),
-
-    /// Internal server error (unexpected failures)
-    Internal(String),
-}
-
-impl From<crate::evidence::EvidenceBundleError> for AppError {
-    fn from(error: crate::evidence::EvidenceBundleError) -> Self {
-        match error {
-            crate::evidence::EvidenceBundleError::InvalidRequest(message) => Self::Bundle(message),
-            crate::evidence::EvidenceBundleError::Conflict(message) => Self::Conflict(message),
-            crate::evidence::EvidenceBundleError::Io(message) => {
-                Self::BundleOutputNotReady(message)
-            }
-        }
-    }
-}
-
-impl AppError {
-    fn quarantine_from_evidence_error(error: crate::evidence::EvidenceBundleError) -> Self {
-        match error {
-            crate::evidence::EvidenceBundleError::InvalidRequest(message) => {
-                Self::Quarantine(message)
-            }
-            crate::evidence::EvidenceBundleError::Conflict(message) => {
-                Self::QuarantineConflict(message)
-            }
-            crate::evidence::EvidenceBundleError::Io(message) => {
-                Self::QuarantineOutputNotReady(message)
-            }
-        }
-    }
-}
-
-impl IntoResponse for AppError {
-    fn into_response(self) -> Response {
-        let (status, code, message, safe_detail, location, next_action) = match self {
-            AppError::Parse(msg) => (
-                StatusCode::BAD_REQUEST,
-                "PARSE_ERROR",
-                msg,
-                "The request message could not be parsed as HL7 v2. Raw message content is not echoed.",
-                Some("message"),
-                "Check the MSH segment, segment terminators, encoding, and mllp_framed setting.",
-            ),
-            // Profile load error is a client error since the profile is provided in the request.
-            // Keep the public message stable and avoid echoing parser detail derived from profile YAML.
-            AppError::ProfileLoad(_) => (
-                StatusCode::BAD_REQUEST,
-                "PROFILE_LOAD_ERROR",
-                PROFILE_LOAD_SAFE_MESSAGE.to_string(),
-                "The supplied inline profile could not be loaded. Raw profile content is not echoed.",
-                Some("profile"),
-                "Run profile lint on the profile, then retry validation with the corrected profile.",
-            ),
-            AppError::Validation(msg) => (
-                StatusCode::BAD_REQUEST,
-                "VALIDATION_ERROR",
-                msg,
-                "The request failed validation before a successful evidence response was produced.",
-                None,
-                "Check request parameters, schema-version fields, and validation issue paths where available.",
-            ),
-            AppError::Redaction(msg) => (
-                StatusCode::BAD_REQUEST,
-                "REDACTION_ERROR",
-                msg,
-                "The redaction policy or redaction run failed before a safe response was produced.",
-                Some("redaction_policy"),
-                "Check safe-analysis policy paths, actions, reasons, and required-field matches before retrying.",
-            ),
-            AppError::BundleOutputNotConfigured => (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "BUNDLE_OUTPUT_NOT_CONFIGURED",
-                "server bundle output root is not configured".to_string(),
-                "The server cannot create evidence bundles until an operator configures a bundle root.",
-                Some("bundle_output_root"),
-                "Configure the server bundle output root and verify readiness before retrying.",
-            ),
-            AppError::BundleOutputNotReady(msg) => (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "BUNDLE_OUTPUT_NOT_READY",
-                msg,
-                "The configured bundle output root is not currently writable or available.",
-                Some("bundle_output_root"),
-                "Check server filesystem permissions and readiness before retrying.",
-            ),
-            AppError::Bundle(msg) => (
-                StatusCode::BAD_REQUEST,
-                "BUNDLE_ERROR",
-                msg,
-                "The bundle request could not be accepted. Server responses use safe bundle identifiers.",
-                Some("bundle_id"),
-                "Use a simple bundle id without path traversal and retry after validating inputs.",
-            ),
-            AppError::Conflict(msg) => (
-                StatusCode::CONFLICT,
-                "BUNDLE_EXISTS",
-                msg,
-                "The requested bundle output already exists under the configured root.",
-                Some("bundle_id"),
-                "Choose a new bundle id or replay the existing bundle instead of overwriting it.",
-            ),
-            AppError::BundleNotFound(msg) => (
-                StatusCode::NOT_FOUND,
-                "BUNDLE_NOT_FOUND",
-                msg,
-                "The requested bundle id was not found under the configured root.",
-                Some("bundle_id"),
-                "Check the bundle id from the bundle creation receipt and retry.",
-            ),
-            AppError::QuarantineOutputNotConfigured => (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "QUARANTINE_OUTPUT_NOT_CONFIGURED",
-                "server quarantine output is enabled but no path is configured".to_string(),
-                "The server cannot write quarantine artifacts until an operator configures a quarantine root.",
-                Some("quarantine.path"),
-                "Configure the quarantine output path or disable quarantine output before retrying.",
-            ),
-            AppError::QuarantineOutputNotReady(msg) => (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "QUARANTINE_OUTPUT_NOT_READY",
-                msg,
-                "The configured quarantine output root is not currently writable or available.",
-                Some("quarantine.path"),
-                "Check server filesystem permissions and readiness before retrying.",
-            ),
-            AppError::Quarantine(msg) => (
-                StatusCode::BAD_REQUEST,
-                "QUARANTINE_ERROR",
-                msg,
-                "The quarantine request could not be written as configured.",
-                Some("quarantine"),
-                "Check quarantine artifact settings and retry with reviewed redaction inputs.",
-            ),
-            AppError::QuarantineConflict(msg) => (
-                StatusCode::CONFLICT,
-                "QUARANTINE_EXISTS",
-                msg,
-                "The generated quarantine output collided with existing output.",
-                Some("quarantine"),
-                "Retry the request or inspect the existing quarantine output before sharing evidence.",
-            ),
-            AppError::Internal(msg) => (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "INTERNAL_ERROR",
-                msg,
-                "The server hit an internal failure. Raw request payloads are not included in this response.",
-                None,
-                "Check server logs and readiness, then retry with the same request only if disclosure policy allows.",
-            ),
-        };
-
-        tracing::warn!(
-            target: "hl7v2_server::evidence",
-            event = audit::EVENT_ERROR,
-            status = status.as_u16(),
-            error_code = code,
-            "request failed"
-        );
-
-        let mut error = ErrorResponse::new(code, message)
-            .with_safe_detail(safe_detail)
-            .with_suggested_next_action(next_action);
-        if let Some(location) = location {
-            error = error.with_location(location);
-        }
-        (status, Json(error)).into_response()
-    }
-}
-
-impl std::fmt::Display for AppError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            AppError::Parse(msg) => write!(f, "Parse error: {}", msg),
-            AppError::ProfileLoad(_) => {
-                write!(f, "Profile load error: {PROFILE_LOAD_SAFE_MESSAGE}")
-            }
-            AppError::Validation(msg) => write!(f, "Validation error: {}", msg),
-            AppError::Redaction(msg) => write!(f, "Redaction error: {}", msg),
-            AppError::BundleOutputNotConfigured => {
-                write!(f, "Bundle output root is not configured")
-            }
-            AppError::BundleOutputNotReady(msg) => {
-                write!(f, "Bundle output root is not ready: {}", msg)
-            }
-            AppError::Bundle(msg) => write!(f, "Bundle error: {}", msg),
-            AppError::Conflict(msg) => write!(f, "Bundle conflict: {}", msg),
-            AppError::BundleNotFound(msg) => write!(f, "Bundle not found: {}", msg),
-            AppError::QuarantineOutputNotConfigured => {
-                write!(f, "Quarantine output path is not configured")
-            }
-            AppError::QuarantineOutputNotReady(msg) => {
-                write!(f, "Quarantine output root is not ready: {}", msg)
-            }
-            AppError::Quarantine(msg) => write!(f, "Quarantine error: {}", msg),
-            AppError::QuarantineConflict(msg) => write!(f, "Quarantine conflict: {}", msg),
-            AppError::Internal(msg) => write!(f, "Internal error: {}", msg),
-        }
-    }
-}
-
-impl From<hl7v2::Error> for AppError {
-    fn from(err: hl7v2::Error) -> Self {
-        AppError::Parse(err.to_string())
-    }
-}
-
-impl From<hl7v2::conformance::profile::ProfileLoadError> for AppError {
-    fn from(_err: hl7v2::conformance::profile::ProfileLoadError) -> Self {
-        AppError::ProfileLoad(PROFILE_LOAD_SAFE_MESSAGE.to_string())
-    }
 }
 
 #[cfg(test)]
